adds scheme to Solarium endpoint
Currently it is not possible to connect to Solr server that is available via HTTPS only. You can pass a scheme to the endpoint to address this, see https://github.com/solariumphp/solarium/issues/267

This commit updates the toSolarium() function to pass the scheme to the endpoint as well. 

Fixes #129