use inverse background color for result_title
In dark theme, the black color was not easily readable.
With additional high contrast mode, the title could not be read at all.

This change uses the inverse background color and should be readable in both cases.
As long as there's only these two themes available, this solution should suffice

Signed-off-by: Tomas Dittmann <chaosmeist3r@gmail.com>